requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dataclasses==0.7
Django==3.0.5
djangorestframework==3.11.0
idna==2.9
pathspec==0.8.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
